+++ core/classes/tpls/app/class.tpl.app.filezilla.php
@@ -70,7 +70,6 @@ class TplAppFilezilla
             TplApp::getActionRun(Action::SWITCH_VERSION, array($neardBins->getFilezilla()->getName(), $version)) . PHP_EOL .
             TplService::getActionCreate(BinFilezilla::SERVICE_NAME) . PHP_EOL .
             TplService::getActionStart(BinFilezilla::SERVICE_NAME) . PHP_EOL .
-            TplAppReload::getActionReload() . PHP_EOL .
             TplApp::getActionExec() . PHP_EOL;
     }